Fix parts of nullability in JSON converter

+++ clients-kt/src/main/kotlin/org/apache/kafka/common/protocol/MessageUtil.kt
@@ -91,14 +91,14 @@ object MessageUtil {
         return value.toShort()
     }
 
-    fun jsonNodeToUnsignedShort(node: JsonNode, about: String): Int {
+    fun jsonNodeToUnsignedShort(node: JsonNode, about: String): UShort {
         val value = jsonNodeToInt(node, about)
 
         if (value < 0 || value > UNSIGNED_SHORT_MAX) throw RuntimeException(
             "$about: value $value does not fit in a 16-bit unsigned integer."
         )
 
-        return value
+        return value.toUShort()
     }
 
     fun jsonNodeToUnsignedInt(node: JsonNode, about: String): Long {
